# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RRHH API",
        "version": "0.1"
    },
    "paths": {
        "/firms": {
            "get": {
                "tags": [
                    "Firm"
                ],
                "summary": "getFirmList",
                "description": "Get all Firms",
                "operationId": "a6d6952c4511607e7f1f6e550bc4d59f",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Firm"
                ],
                "summary": "createFirm",
                "description": "Create Firm",
                "operationId": "b55b8bd693ff626db705506b09066b0b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "desc",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/firms/{id}": {
            "get": {
                "tags": [
                    "Firm"
                ],
                "summary": "getFirmItem",
                "description": "Get Firm",
                "operationId": "e3005e60f1c479cf040b3787d37d8a3e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Firm",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Firm"
                ],
                "summary": "updateFirm",
                "description": "Update Firm",
                "operationId": "7fcda2a05724b40e17da2dc84ef787fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Firm",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "desc",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Firm"
                ],
                "summary": "deleteFirm",
                "description": "Delete Firm",
                "operationId": "fe2fb8e99d6ef8d48902ab20ce7c5cf8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Firm",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "in": "header"
            }
        }
    }
}